pyproject.toml
[project]
name = "ai-agent-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
